fix: refactored reference_editor tools to separated files and cleaned the code
