+++ agentic-backend/agentic_backend/agents/reference_editor/__init__.py
@@ -1,0 +1,19 @@
+# Copyright Thales 2025
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Reference Editor Agent for generating reference PowerPoint and Word documents."""
+
+from agentic_backend.agents.reference_editor.reference_editor import ReferenceEditor
+
+__all__ = ["ReferenceEditor"]
