+++ agentic-backend/agentic_backend/agents/reference_editor/reference_editor.py
@@ -11,31 +11,23 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""Reference Editor Agent for generating reference PowerPoint and Word documents."""
+
 from __future__ import annotations
 
 import logging
-import tempfile
-import uuid
-from pathlib import Path
 
 from fred_core import OwnerFilter
-from jsonschema import Draft7Validator
 from langchain.agents import create_agent
-from langchain.tools import tool
 from langgraph.graph.state import CompiledStateGraph
 from langgraph.types import Checkpointer
 
-from agentic_backend.agents.reference_editor.powerpoint_template_util import (
-    fill_slide_from_structured_response_async,
-    fill_word_from_structured_response_async,
-    referenceSchema,
-)
+from agentic_backend.agents.reference_editor.export_tools import ExportTools
+from agentic_backend.agents.reference_editor.validation_tools import ValidationTools
 from agentic_backend.application_context import get_default_chat_model
-from agentic_backend.common.kf_base_client import KfBaseClient
-from agentic_backend.common.kf_vectorsearch_client import (
-    VectorSearchClient,
-)
 from agentic_backend.common.mcp_runtime import MCPRuntime
+from agentic_backend.common.structures import AgentChatOptions
 from agentic_backend.core.agents.agent_flow import AgentFlow
 from agentic_backend.core.agents.agent_spec import (
     AgentTuning,
@@ -49,84 +41,47 @@ from agentic_backend.core.agents.runtime_context import (
     get_document_uids,
     get_vector_search_scopes,
 )
-from agentic_backend.core.chatbot.chat_schema import (
-    LinkKind,
-    LinkPart,
-)
+from agentic_backend.core.runtime_source import expose_runtime_source
 
 logger = logging.getLogger(__name__)
-EXPECTED_REFERENCE_SECTIONS = {"informationsProjet", "contexte", "syntheseProjet"}
 
 
-def _normalize_reference_payload_for_validation(
-    data: dict | None,
-) -> tuple[dict | None, str | None]:
-    """
-    Accept both payload styles:
-    - {"data": {...}}  (preferred, matches tool schema)
-    - {...}            (legacy convenience)
-    """
-    if data is None:
-        return (
-            None,
-            "Missing required argument: data. Call validator_tool(data={...}) with the structured payload.",
-        )
-    if not isinstance(data, dict):
-        return None, "Invalid payload type. Expected a JSON object."
+@expose_runtime_source("agent.ReferenceEditor")
+class ReferenceEditor(AgentFlow):
+    """Agent that generates reference PowerPoint and Word documents from extracted data."""
 
-    payload = data.get("data", data)
-    if not isinstance(payload, dict):
-        return (
-            None,
-            "Invalid payload format. Expected `data` to be a JSON object.",
-        )
-
-    keys = set(payload.keys())
-    if keys != EXPECTED_REFERENCE_SECTIONS:
-        return (
-            None,
-            "Bad root key format. Expected sections are: "
-            '{"informationsProjet": {...}, "contexte": {...}, "syntheseProjet": {...}} '
-            "(inside `data` if wrapped).",
-        )
-
-    return payload, None
-
-
-# --- Configuration & Tuning ---
-# ------------------------------
-TUNING = AgentTuning(
-    role="Reference Editor",
-    description="Extracts information from reference powerpoint to fill another given reference PowerPoint template.",
-    mcp_servers=[MCPServerRef(id="mcp-knowledge-flow-mcp-text")],
-    tags=[],
-    fields=[
-        FieldSpec(
-            key="ppt.template_key",
-            type="text",
-            title="PowerPoint Template Key",
-            description="Agent asset key for the .pptx template.",
-            ui=UIHints(group="PowerPoint"),
-            default="ref_template.pptx",
-        ),
-        FieldSpec(
-            key="word.template_key",
-            type="text",
-            title="Word Template Key",
-            description="Agent asset key for the .docx template.",
-            ui=UIHints(group="Word"),
-            default="ref_template.docx",
-        ),
-        FieldSpec(
-            key="prompts.system",
-            type="prompt",
-            title="System Prompt",
-            description=(
-                "High-level instructions for the agent. "
-                "State the mission, how to use the available tools, and constraints."
+    tuning = AgentTuning(
+        role="Reference Editor",
+        description="Extracts information from reference powerpoint to fill another given reference PowerPoint template.",
+        mcp_servers=[MCPServerRef(id="mcp-knowledge-flow-mcp-text")],
+        tags=["document", "powerpoint", "reference"],
+        fields=[
+            FieldSpec(
+                key="ppt.template_key",
+                type="text",
+                title="PowerPoint Template Key",
+                description="Agent asset key for the .pptx template.",
+                ui=UIHints(group="PowerPoint"),
+                default="ref_template.pptx",
             ),
-            required=True,
-            default="""
+            FieldSpec(
+                key="word.template_key",
+                type="text",
+                title="Word Template Key",
+                description="Agent asset key for the .docx template.",
+                ui=UIHints(group="Word"),
+                default="ref_template.docx",
+            ),
+            FieldSpec(
+                key="prompts.system",
+                type="prompt",
+                title="System Prompt",
+                description=(
+                    "High-level instructions for the agent. "
+                    "State the mission, how to use the available tools, and constraints."
+                ),
+                required=True,
+                default="""
 Tu es un agent spécialisé dans l'extraction d'informations structurées depuis des documents PowerPoint de référence pour remplir un template PowerPoint standardisé.
 
 # MISSION
@@ -275,6 +230,12 @@ Autres règles :
 - À chaque génération ou modification : validation + templetisation complète (les deux dans le même tour)
 - Le lien de téléchargement est automatiquement généré par l'outil de templetisation, utilise-le directement
 
+# RÈGLE DE RESTITUTION DES TÉLÉCHARGEMENTS (OBLIGATOIRE)
+- Si template_tool ou word_template_tool retourne un LinkPart, ne le réécris jamais en texte.
+- N'affiche jamais d'URL brute, de markdown `[Download ...]`, ni de ligne `Download ...`.
+- Laisse uniquement le LinkPart pour le téléchargement.
+- Tu peux dire en texte : "Le bouton de téléchargement est ci-dessous."
+
 # EXEMPLE DE RÉPONSE UTILISATEUR
 
 J'ai extrait les informations depuis les documents de référence et généré votre document de référence.
@@ -289,29 +250,52 @@ J'ai extrait les informations depuis les documents de référence et généré v
 [Lien de téléchargement retourné par l'outil de templetisation]
 
 """,
-            ui=UIHints(group="Prompts", multiline=True, markdown=True),
-        ),
-    ],
-)
+                ui=UIHints(group="Prompts", multiline=True, markdown=True),
+            ),
+            FieldSpec(
+                key="chat_options.attach_files",
+                type="boolean",
+                title="Allow file attachments",
+                description="Show file upload/attachment controls for this agent.",
+                required=False,
+                default=True,
+                ui=UIHints(group="Chat options"),
+            ),
+            FieldSpec(
+                key="chat_options.libraries_selection",
+                type="boolean",
+                title="Document libraries picker",
+                description="Let users select document libraries/knowledge sources for this agent.",
+                required=False,
+                default=True,
+                ui=UIHints(group="Chat options"),
+            ),
+        ],
+    )
 
-
-class ReferenceEditor(AgentFlow):
-    """
-    Simplified agent to generate a PowerPoint slide with LLM content
-    and return a structured download link.
-    """
-
-    tuning = TUNING
+    default_chat_options = AgentChatOptions(
+        attach_files=True,
+        libraries_selection=True,
+        search_rag_scoping=False,
+        search_policy_selection=False,
+        deep_search_delegate=False,
+    )
 
     async def async_init(self, runtime_context: RuntimeContext):
-        await super().async_init(runtime_context)
+        """Initialize agent and tool helpers."""
+        await super().async_init(runtime_context=runtime_context)
         self.mcp = MCPRuntime(agent=self)
         await self.mcp.init()
 
+        # Initialize tool helpers
+        self.export_tools = ExportTools(self)
+        self.validation_tools = ValidationTools(self)
+
     async def aclose(self):
+        """Clean up resources."""
         await self.mcp.aclose()
 
-    def _build_vector_search_scope_options(self) -> dict:
+    def build_vector_search_scope_options(self) -> dict:
         """
         Build strict retrieval constraints for direct vector-search calls done inside
         template/image utilities.
@@ -337,250 +321,18 @@ class ReferenceEditor(AgentFlow):
     def get_compiled_graph(
         self, checkpointer: Checkpointer | None = None
     ) -> CompiledStateGraph:
-        template_tool = self.get_template_tool()
-        word_template_tool = self.get_word_template_tool()
-        validator_tool = self.get_validator_tool()
-        base_system_prompt = self.render(self.get_tuned_text("prompts.system") or "")
-        download_guardrail = """
-# RÈGLE DE RESTITUTION DES TÉLÉCHARGEMENTS (OBLIGATOIRE)
-- Si template_tool ou word_template_tool retourne un LinkPart, ne le réécris jamais en texte.
-- N'affiche jamais d'URL brute, de markdown `[Download ...]`, ni de ligne `Download ...`.
-- Laisse uniquement le LinkPart pour le téléchargement.
-- Tu peux dire en texte : "Le bouton de téléchargement est ci-dessous."
-"""
-        system_prompt = (
-            f"{base_system_prompt.rstrip()}\n\n{download_guardrail.strip()}"
-            if base_system_prompt
-            else download_guardrail.strip()
-        )
-
+        """Create the agent graph with all tools."""
         return create_agent(
             model=get_default_chat_model(),
-            system_prompt=system_prompt,
+            system_prompt=self.render(self.get_tuned_text("prompts.system") or ""),
             tools=[
-                template_tool,
-                word_template_tool,
-                validator_tool,
+                # Validation
+                self.validation_tools.get_validator_tool(),
+                # Export
+                self.export_tools.get_ppt_template_tool(),
+                self.export_tools.get_word_template_tool(),
+                # MCP tools for RAG search
                 *self.mcp.get_tools(),
             ],
             checkpointer=checkpointer,
         )
-
-    def get_validator_tool(self):
-        @tool
-        async def validator_tool(data: dict | None = None):
-            """
-            Outil permettant de valider le format des données avant de les passer à l'outil de templetisation.
-            L'outil retourne [] si le schéma est valide et la liste des erreurs sinon.
-
-            IMPORTANT : Si cet outil retourne [] (liste vide), tu DOIS IMMÉDIATEMENT appeler template_tool(data={{...}})
-            avec exactement les mêmes données dans le MÊME tour de conversation. Ne t'arrête pas ici.
-            """
-            payload, error_message = _normalize_reference_payload_for_validation(data)
-            if error_message:
-                return error_message
-
-            def shorten_error_message(error):
-                """Convert verbose validation errors to concise messages"""
-                field_path = ".".join(str(p) for p in error.path) or "root"
-                if error.validator == "type":
-                    return f"{field_path} type invalid. Expected {error.schema.get('type')}."
-                if error.validator == "required":
-                    return f"{field_path} missing required field."
-                return f"{field_path} invalid. Reason: {error.validator}."
-
-            validator = Draft7Validator(referenceSchema)
-
-            errors = [shorten_error_message(e) for e in validator.iter_errors(payload)]
-            if not errors:
-                return "✓ Validation réussie ! Appelle maintenant template_tool(data={...}) avec ces mêmes données."
-            return errors
-
-        return validator_tool
-
-    def get_template_tool(self):
-        tool_schema = {
-            "type": "object",
-            "properties": {
-                "data": referenceSchema,  # todo: get it by parsing a tuning field
-            },
-            "required": ["data"],
-        }
-
-        @tool(args_schema=tool_schema)
-        async def template_tool(data: dict):
-            """
-            Outil permettant de templétiser le fichier envoyé par l'utilisateur.
-            La nature du fichier importe peu tant que le format des données est respecté. Tu n'as pas besoin de préciser quel fichier,
-            l'outil possède déjà cette information.
-            L'outil retourne un LinkPart pour l'interface. Ne jamais réécrire ce lien en texte/Markdown.
-            """
-            # 1. Fetch template from secure asset storage
-            template_key = (
-                self.get_tuned_text("ppt.template_key") or "simple_template.pptx"
-            )
-            template_path = await self.fetch_config_blob_to_tempfile(
-                template_key, suffix=".pptx"
-            )
-
-            # 2. Save the modified presentation to a temp file
-            with tempfile.NamedTemporaryFile(
-                delete=False, suffix=".pptx", prefix="result_"
-            ) as out:
-                output_path = Path(out.name)
-                # Extract the actual data from the wrapper
-                actual_data = data.get("data", data)
-
-                vector_search_client = VectorSearchClient(agent=self)
-                search_options = self._build_vector_search_scope_options()
-                kf_base_client = KfBaseClient(
-                    allowed_methods=frozenset({"GET", "POST"}), agent=self
-                )
-                await fill_slide_from_structured_response_async(
-                    template_path,
-                    actual_data,
-                    output_path,
-                    vector_search_client,
-                    kf_base_client,
-                    search_options=search_options,
-                )
-
-            # 3. Upload the generated asset to user storage
-            import asyncio
-
-            # user_id_to_store_asset = self.get_end_user_id()
-            # Use UUID to generate a unique filename that won't trigger versioning conflicts
-            unique_id = str(uuid.uuid4())
-            final_key = f"Generated_Slide_{unique_id}.pptx"
-
-            # Try upload with retry logic in case of temporary backend issues
-            max_retries = 2
-            upload_succeeded = False
-            last_error = None
-
-            for attempt in range(max_retries):
-                try:
-                    with open(output_path, "rb") as f_out:
-                        upload_result = await self.upload_user_blob(
-                            key=final_key,
-                            file_content=f_out,
-                            filename=final_key,  # Use the same unique filename to avoid versioning conflicts
-                            content_type="application/vnd.openxmlformats-officedocument.presentationml.presentation",
-                        )
-                    upload_succeeded = True
-                    break  # Success, exit retry loop
-                except Exception as e:
-                    last_error = e
-                    if attempt < max_retries - 1:
-                        # Wait a bit before retrying
-                        await asyncio.sleep(1)
-                        continue
-
-            if upload_succeeded:
-                # 4. Construct the structured message for the UI
-                final_download_url = upload_result.download_url
-
-                return LinkPart(
-                    href=final_download_url,
-                    title=f"Download {upload_result.file_name}",
-                    kind=LinkKind.download,
-                    mime="application/vnd.openxmlformats-officedocument.presentationml.presentation",
-                ).model_dump(mode="json")
-            else:
-                return last_error
-
-        return template_tool
-
-    def get_word_template_tool(self):
-        tool_schema = {
-            "type": "object",
-            "properties": {
-                "data": referenceSchema,  # todo: get it by parsing a tuning field
-            },
-            "required": ["data"],
-        }
-
-        @tool(args_schema=tool_schema)
-        async def word_template_tool(data: dict):
-            """
-            Outil permettant de templétiser un fichier Word envoyé par l'utilisateur.
-            La nature du fichier importe peu tant que le format des données est respecté. Tu n'as pas besoin de préciser quel fichier,
-            l'outil possède déjà cette information.
-            L'outil retourne un LinkPart pour l'interface. Ne jamais réécrire ce lien en texte/Markdown.
-            """
-            # 1. Fetch template from secure asset storage
-            template_key = (
-                self.get_tuned_text("word.template_key") or "simple_template.docx"
-            )
-            template_path = await self.fetch_config_blob_to_tempfile(
-                template_key, suffix=".docx"
-            )
-
-            # 2. Save the modified document to a temp file
-            with tempfile.NamedTemporaryFile(
-                delete=False, suffix=".docx", prefix="result_"
-            ) as out:
-                output_path = Path(out.name)
-                # Extract the actual data from the wrapper
-                actual_data = data.get("data", data)
-                # Create clients for image search
-                from agentic_backend.common.kf_base_client import KfBaseClient
-                from agentic_backend.common.kf_vectorsearch_client import (
-                    VectorSearchClient,
-                )
-
-                vector_search_client = VectorSearchClient(agent=self)
-                search_options = self._build_vector_search_scope_options()
-                kf_base_client = KfBaseClient(
-                    allowed_methods=frozenset({"GET", "POST"}), agent=self
-                )
-                await fill_word_from_structured_response_async(
-                    template_path,
-                    actual_data,
-                    output_path,
-                    vector_search_client,
-                    kf_base_client,
-                    search_options=search_options,
-                )
-
-            # 3. Upload the generated asset to user storage
-            import asyncio
-            import uuid
-
-            # user_id_to_store_asset = self.get_end_user_id()
-            # Use UUID to generate a unique filename that won't trigger versioning conflicts
-            unique_id = str(uuid.uuid4())
-            final_key = f"Generated_Document_{unique_id}.docx"
-
-            # Try upload with retry logic in case of temporary backend issues
-            max_retries = 2
-            for attempt in range(max_retries):
-                try:
-                    with open(output_path, "rb") as f_out:
-                        upload_result = await self.upload_user_blob(
-                            key=final_key,
-                            file_content=f_out,
-                            filename=final_key,  # Use the same unique filename to avoid versioning conflicts
-                            content_type="application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-                        )
-                    break  # Success, exit retry loop
-                except Exception:
-                    if attempt < max_retries - 1:
-                        # Wait a bit before retrying
-                        await asyncio.sleep(1)
-                        continue
-                    else:
-                        # Final attempt failed, re-raise
-                        raise
-
-            # 4. Construct the structured message for the UI
-            final_download_url = upload_result.download_url
-
-            return LinkPart(
-                href=final_download_url,
-                title=f"Download {upload_result.file_name}",
-                kind=LinkKind.download,
-                mime="application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-            ).model_dump(mode="json")
-
-        return word_template_tool
